[ICU-900] Fixed some compiler warnings. X-Commit-URL: https://ssl.icu-project.org/trac/changeset/8681 Former-commit-id: d52216eca9759f6114fa95a215b7564f9928607e

+++ icu4c/source/extra/uconv/uwmsg.c
@@ -230,7 +230,7 @@ U_CAPI const UChar *u_wmsg_errorName(UErrorCode err)
             textMsg = error;
         }
         msg = (UChar*)malloc((strlen(textMsg)+1)*sizeof(msg[0]));
-        u_charsToUChars(textMsg, msg, strlen(textMsg)+1);
+        u_charsToUChars(textMsg, msg, (int32_t)(strlen(textMsg)+1));
     }
 
     if(err>=0)
